re-organize util imports in localstack.services package
