+++ localstack/services/plugins.py
@@ -14,7 +14,9 @@ from requests.models import Request
 from localstack import config
 from localstack.config import ServiceProviderConfig
 from localstack.utils.bootstrap import get_enabled_apis, is_api_enabled, log_duration
-from localstack.utils.common import call_safe, poll_condition, wait_for_port_status
+from localstack.utils.functions import call_safe
+from localstack.utils.net import wait_for_port_status
+from localstack.utils.sync import poll_condition
 
 # set up logger
 LOG = logging.getLogger(__name__)
